feat: change inequality parser

+++ src/main/java/com/example/mtg_deckbuilder/utils/Inequality.java
@@ -1,3 +1,3 @@
 package com.example.mtg_deckbuilder.utils;
 
-public record Inequality(String operator, Double value) {}
+public record Inequality(String operator, Integer value) {}
